Sea: split out sea fleet from eval
